check if module already installed

+++ local-cookbooks/development/recipes/npm.rb
@@ -4,5 +4,6 @@ node.development.application.npm_packages.each_pair do |pkg, ver|
         code <<-EOF
             npm install -g #{pkg}@#{ver}
         EOF
+        not_if { ::File.exists?("/usr/local/lib/node_modules/#{pkg}") }
     end
 end
